[Math] Correctly use floating-point value in tt-generated vector Lerp()

+++ Molten.Math/Vectors/Vector2D.cs
@@ -131,16 +131,15 @@ namespace Molten.Math
         /// <param name="start">The start vector.</param>
         /// <param name="end">The end vector.</param>
         /// <param name="amount">Value between 0 and 1 indicating the weight of <paramref name="end"/>.</param>
-        /// <param name="result">When the method completes, contains the linear interpolation of the two vectors.</param>
         /// <remarks>
         /// Passing <paramref name="amount"/> a value of 0 will cause <paramref name="start"/> to be returned; a value of 1 will cause <paramref name="end"/> to be returned. 
         /// </remarks>
-        public static Vector2D Lerp(ref Vector2D start, ref Vector2D end, float amount)
+        public static Vector2D Lerp(ref Vector2D start, ref Vector2D end, double amount)
         {
 			return new Vector2D()
 			{
-				X = (double)((1f - amount) * start.X + amount * end.X),
-				Y = (double)((1f - amount) * start.Y + amount * end.Y),
+				X = (double)((1D - amount) * start.X + amount * end.X),
+				Y = (double)((1D - amount) * start.Y + amount * end.Y),
 			};
         }
