+++ Molten.Math/Vectors/Vector3D.cs
@@ -144,17 +144,16 @@ namespace Molten.Math
         /// <param name="start">The start vector.</param>
         /// <param name="end">The end vector.</param>
         /// <param name="amount">Value between 0 and 1 indicating the weight of <paramref name="end"/>.</param>
-        /// <param name="result">When the method completes, contains the linear interpolation of the two vectors.</param>
         /// <remarks>
         /// Passing <paramref name="amount"/> a value of 0 will cause <paramref name="start"/> to be returned; a value of 1 will cause <paramref name="end"/> to be returned. 
         /// </remarks>
-        public static Vector3D Lerp(ref Vector3D start, ref Vector3D end, float amount)
+        public static Vector3D Lerp(ref Vector3D start, ref Vector3D end, double amount)
         {
 			return new Vector3D()
 			{
-				X = (double)((1f - amount) * start.X + amount * end.X),
-				Y = (double)((1f - amount) * start.Y + amount * end.Y),
-				Z = (double)((1f - amount) * start.Z + amount * end.Z),
+				X = (double)((1D - amount) * start.X + amount * end.X),
+				Y = (double)((1D - amount) * start.Y + amount * end.Y),
+				Z = (double)((1D - amount) * start.Z + amount * end.Z),
 			};
         }
