+++ Molten.Math/Vectors/Vector3UL.cs
@@ -144,17 +144,16 @@ namespace Molten.Math
         /// <param name="start">The start vector.</param>
         /// <param name="end">The end vector.</param>
         /// <param name="amount">Value between 0 and 1 indicating the weight of <paramref name="end"/>.</param>
-        /// <param name="result">When the method completes, contains the linear interpolation of the two vectors.</param>
         /// <remarks>
         /// Passing <paramref name="amount"/> a value of 0 will cause <paramref name="start"/> to be returned; a value of 1 will cause <paramref name="end"/> to be returned. 
         /// </remarks>
-        public static Vector3UL Lerp(ref Vector3UL start, ref Vector3UL end, float amount)
+        public static Vector3UL Lerp(ref Vector3UL start, ref Vector3UL end, double amount)
         {
 			return new Vector3UL()
 			{
-				X = (ulong)((1f - amount) * start.X + amount * end.X),
-				Y = (ulong)((1f - amount) * start.Y + amount * end.Y),
-				Z = (ulong)((1f - amount) * start.Z + amount * end.Z),
+				X = (ulong)((1D - amount) * start.X + amount * end.X),
+				Y = (ulong)((1D - amount) * start.Y + amount * end.Y),
+				Z = (ulong)((1D - amount) * start.Z + amount * end.Z),
 			};
         }
