+++ Molten.Math/Vectors/Vector4D.cs
@@ -157,18 +157,17 @@ namespace Molten.Math
         /// <param name="start">The start vector.</param>
         /// <param name="end">The end vector.</param>
         /// <param name="amount">Value between 0 and 1 indicating the weight of <paramref name="end"/>.</param>
-        /// <param name="result">When the method completes, contains the linear interpolation of the two vectors.</param>
         /// <remarks>
         /// Passing <paramref name="amount"/> a value of 0 will cause <paramref name="start"/> to be returned; a value of 1 will cause <paramref name="end"/> to be returned. 
         /// </remarks>
-        public static Vector4D Lerp(ref Vector4D start, ref Vector4D end, float amount)
+        public static Vector4D Lerp(ref Vector4D start, ref Vector4D end, double amount)
         {
 			return new Vector4D()
 			{
-				X = (double)((1f - amount) * start.X + amount * end.X),
-				Y = (double)((1f - amount) * start.Y + amount * end.Y),
-				Z = (double)((1f - amount) * start.Z + amount * end.Z),
-				W = (double)((1f - amount) * start.W + amount * end.W),
+				X = (double)((1D - amount) * start.X + amount * end.X),
+				Y = (double)((1D - amount) * start.Y + amount * end.Y),
+				Z = (double)((1D - amount) * start.Z + amount * end.Z),
+				W = (double)((1D - amount) * start.W + amount * end.W),
 			};
         }
